Updated code
Added parts for the hpc server.
Added part to put distance_df in long format for easier vizualisations and added part to create graphs for single species per location is occurence in.
Slight changes in names of functions.R script so that these matches in the AlienDetective_2.R script and fixed the y-axes.